+++ sndata/loss/__init__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# -*- coding: UTF-8 -*-
+
+"""The ``loss`` module provides access to supernova data from the Lick
+Observatory Supernova Search (LOSS). For the spectroscopic compliment to this
+survey, see the ``bsnip`` module.
+"""
+
+from ._ganeshalingam13 import Ganeshalingam13
+
+survey_name = 'Lick Observatory Supernova Search'
+survey_abbrev = 'LOSS'