+++ sndata/sdss/_sako18spec.py
@@ -184,7 +184,7 @@ class Sako18Spec(SpectroscopicRelease):
         for file_name in self._table_names:
             utils.download_file(
                 url=self._base_url + file_name,
-                path=self._table_dir / file_name,
+                destination=self._table_dir / file_name,
                 force=force,
                 timeout=timeout
             )